update DW_OP_overlay to check for width going beyond overlay
The definition of DW_OP_overlay requires that the overlay does not
extend outside of the bounds of the storage of the overlay location.
In other words, the width must be less than or equal to the size of
the overlay location minus the overlay location's offset.

Handle this error case and update/add examples.

Signed-off-by: Tankut Baris Aktemur <tankut.baris.aktemur@intel.com>