Add web-based playground in js/
Use js_of_ocaml and some ppx-based syntax extensions to derive a
web-based playground for Dwarf_locstack.

Minimal changes were required to the original module source, the biggest
being a rename from dwarf-locstack to dwarf_locstack to make it a valid
module identifier. The remaining changes are just to satisfy some
warnings that default to errors. I masked out some of these with
compiler options, but fixed a couple in the source.